+++ src/Menu.java
@@ -1,31 +1,20 @@
 
 import javax.swing.JOptionPane;
 import Objects.*;
-import java.io.*;
-import data_structures.*;
+import data_structures.DoubleNode;
 
 public class Menu {
 
-    public DoubleList listaEmpleados;
-
-    public Menu() {
-        this.listaEmpleados = new DoubleList();
+    ListaEmpleados listaEmpleados;
+    Usuario user;
+    
+    public Menu(ListaEmpleados lista) {
+        this.listaEmpleados = lista;
     }
 
-    // sets y gets
-    public DoubleList getlistaEmpleados() {
-        return listaEmpleados;
-    }
-
-    public void setlistaEmpleados(DoubleList listaEmpleados) {
-        this.listaEmpleados = listaEmpleados;
-    }
-
-    // metodos propios de la clase
-
+    // error en el menu empleado case 1
     public void menuEmpleado() {
-        // Menú empleado bandeja
-        Empleado meb = new Empleado();
+        Empleado meb = new Empleado(user);
         int opcion;
         do {
             String input = JOptionPane.showInputDialog(
@@ -35,17 +24,17 @@ public class Menu {
                             "4. Ver borradores\n" +
                             "5. Salir\n" +
                             "Seleccione una opción:");
-
+                            
             opcion = Integer.parseInt(input);
 
             switch (opcion) {
                 case 1:
-                    meb.bandeja.redactarMensaje();
+                meb.bandeja.redactarMensaje();
                     break;
                 case 2:
                     meb.revisarBandeja();
                     break;
-                case 3:
+                    case 3:
                     meb.verMensajesLeidos();
                     break;
                 case 4:
@@ -59,15 +48,16 @@ public class Menu {
                     break;
             }
         } while (opcion != 5);
+        System.exit(0);
     }
-
+    
     public void menuAdmin() {
         // Menú administrador bandeja
-        Administrador mab = new Administrador();
+        Administrador mab = new Administrador(listaEmpleados, user);
         int opcion;
         do {
             String input = JOptionPane.showInputDialog(
-                    "1. Registrar usuario\n" +
+                "1. Registrar usuario\n" +
                             "2. Cambiar contraseña\n" +
                             "3. Eliminar usuario\n" +
                             "4. Actualizar información\n" +
@@ -80,7 +70,7 @@ public class Menu {
                 case 1:
                     mab.registrarUsuarios();
                     break;
-                case 2:
+                    case 2:
                     String userIDInput = JOptionPane.showInputDialog("Ingrese el ID del usuario:");
                     Long userID = Long.parseLong(userIDInput);
 
@@ -104,48 +94,31 @@ public class Menu {
                 case 5:
                     JOptionPane.showMessageDialog(null, "Hasta luego!");
                     break;
-                default:
+                    default:
                     JOptionPane.showMessageDialog(null, "Opción inválida");
                     break;
-            }
+                }
         } while (opcion != 5);
+        System.exit(0);
     }
-
     
     public void Login() {
 
-        // lee un usuario y una contraseña
         Long idLogin = Long.parseLong(JOptionPane.showInputDialog(null, "Usuario", "iniciar sesion", 1));
         String LoginPassword = JOptionPane.showInputDialog(null, "Contraseña", "iniciar sesion", 1);
 
-        // lee el archivo de contraseñas y valida su identidad
-        File file = new File((System.getProperty("user.dir") + "/src/Files/Password.txt"));
-;
-        try (BufferedReader br = new BufferedReader(new FileReader(file))) {
-            String line;
-            while ((line = br.readLine()) != null) {
-
-                String[] data = line.split(" ");
-                ValidateUser(idLogin, LoginPassword, data);
-
+        DoubleNode actUser = listaEmpleados.getLista().first(); 
+        while (actUser != null) {
+            if((((Empleado) actUser.getDato()).getUser().getId().equals(idLogin)) && ((Empleado) actUser.getDato()).getUser().getContrasena().equals(LoginPassword)){
+                if (((Empleado) actUser.getDato()).getUser().getTipoUsuario().equals("administrador")) {
+                    menuAdmin();
+                    this.user = ((Empleado)actUser.getDato()).getUser();
+                } else {
+                    menuEmpleado();
+                    this.user = ((Empleado)actUser.getDato()).getUser();
+                }
             }
-        } catch (Exception e) {
-            System.out.println("error al leer las credenciales");
-        }
-    }
-    private void ValidateUser(Long idLogin, String LoginPassword, String[] data) {
-        Long idUser = Long.parseLong(data[0]);
-        String passwd = data[1];
-        String tipoUsuario = data[2];
-        if (idLogin == idUser && LoginPassword.equals(passwd)) {
-            if (tipoUsuario.equals("administrador")) {
-                // si es verdadero es Admin
-                menuAdmin();
-            } else {
-                // si es falso es empleado
-                menuEmpleado();
-            }
+            actUser.getNext();
         }
     }
 }
-